+++ js/listener.js
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2020
+ * (c) Copyright Ascensio System SIA 2021
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -87,16 +87,12 @@
     };
 
     window.addEventListener("message", function (event) {
-        if ($(OCA.Onlyoffice.frameSelector)[0].contentWindow !== event.source
+        if (!$(OCA.Onlyoffice.frameSelector).length
+            || $(OCA.Onlyoffice.frameSelector)[0].contentWindow !== event.source
             || !event.data["method"]) {
             return;
         }
         switch (event.data.method) {
-            case "editorShowHeaderButton":
-                if (OCA.Onlyoffice.ShowHeaderButton) {
-                    OCA.Onlyoffice.ShowHeaderButton();
-                }
-                break;
             case "editorRequestClose":
                 OCA.Onlyoffice.onRequestClose();
                 break;
@@ -122,5 +118,11 @@
                 break;
         }
     }, false);
+
+    window.addEventListener("popstate", function (event) {
+        if ($(OCA.Onlyoffice.frameSelector).length) {
+            OCA.Onlyoffice.onRequestClose();
+        }
+    });
 
 })(OCA);