+++ js/main.js
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2020
+ * (c) Copyright Ascensio System SIA 2021
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -27,7 +27,7 @@
 
     OCA.Onlyoffice.setting = {};
 
-    OCA.Onlyoffice.CreateFile = function (name, fileList) {
+    OCA.Onlyoffice.CreateFile = function (name, fileList, templateId) {
         var dir = fileList.getCurrentDirectory();
 
         if (!OCA.Onlyoffice.setting.sameTab || OCA.Onlyoffice.Desktop) {
@@ -39,6 +39,10 @@
             name: name,
             dir: dir
         };
+
+        if (templateId) {
+            createData.templateId = templateId;
+        }
 
         if ($("#isPublic").val()) {
             createData.shareToken = encodeURIComponent($("#sharingToken").val());
@@ -102,7 +106,9 @@
 
             $("body").addClass("onlyoffice-inline");
 
-            OCA.Files.Sidebar.close();
+            if (OCA.Files.Sidebar) {
+                OCA.Files.Sidebar.close();
+            }
 
             var scrollTop = $(window).scrollTop();
             $(OCA.Onlyoffice.frameSelector).css("top", scrollTop);
@@ -112,38 +118,10 @@
         }
     };
 
-    OCA.Onlyoffice.ShowHeaderButton = function () {
-        if ($("#onlyofficeHeader").length) {
-            return;
-        }
-
-        var wrapper = $("<div id='onlyofficeHeader' />")
-
-        var btnClose = $("<a class='icon icon-close-white'></a>");
-        btnClose.on("click", function() {
-            OCA.Onlyoffice.onRequestClose();
-        });
-        wrapper.prepend(btnClose);
-
-        if (!$("#isPublic").val()) {
-            var btnShare = $("<a class='icon icon-shared icon-white'></a>");
-            btnShare.on("click", function () {
-                OCA.Onlyoffice.OpenShareDialog();
-            })
-            wrapper.prepend(btnShare);
-        }
-
-        if (!$("#header .header-right").length) {
-            $("#header").append("<div class='header-right'></div>");
-        }
-        wrapper.prependTo(".header-right");
-    };
-
     OCA.Onlyoffice.CloseEditor = function () {
         OCA.Onlyoffice.frameSelector = null;
 
         $("body").removeClass("onlyoffice-inline");
-        $("#onlyofficeHeader").remove();
 
         OCA.Onlyoffice.context = null;
 
@@ -158,7 +136,7 @@
 
     OCA.Onlyoffice.OpenShareDialog = function () {
         if (OCA.Onlyoffice.context) {
-            if (!$("#app-sidebar").is(":visible")) {
+            if (!$("#app-sidebar, #app-sidebar-vue").is(":visible")) {
                 OCA.Onlyoffice.context.fileList.showDetailsView(OCA.Onlyoffice.context.fileName, "sharing");
                 OC.Apps.showAppSidebar();
             } else {
@@ -234,7 +212,7 @@
                     if (!config.mime) {
                         return true;
                     }
-                    fileList.fileActions.registerAction({
+                    OCA.Files.fileActions.registerAction({
                         name: "onlyofficeOpen",
                         displayName: t(OCA.Onlyoffice.AppName, "Open in ONLYOFFICE"),
                         mime: config.mime,
@@ -244,11 +222,11 @@
                     });
 
                     if (config.def) {
-                        fileList.fileActions.setDefault(config.mime, "onlyofficeOpen");
+                        OCA.Files.fileActions.setDefault(config.mime, "onlyofficeOpen");
                     }
 
                     if (config.conv) {
-                        fileList.fileActions.registerAction({
+                        OCA.Files.fileActions.registerAction({
                             name: "onlyofficeConvert",
                             displayName: t(OCA.Onlyoffice.AppName, "Convert with ONLYOFFICE"),
                             mime: config.mime,
@@ -279,7 +257,11 @@
                 iconClass: "icon-onlyoffice-new-docx",
                 fileType: "docx",
                 actionHandler: function (name) {
-                    OCA.Onlyoffice.CreateFile(name + ".docx", fileList);
+                    if (!$("#isPublic").val() && OCA.Onlyoffice.TemplateExist("document")) {
+                        OCA.Onlyoffice.OpenTemplatePicker(name, ".docx", "document");
+                    } else {
+                        OCA.Onlyoffice.CreateFile(name + ".docx", fileList);
+                    }
                 }
             });
 
@@ -290,7 +272,11 @@
                 iconClass: "icon-onlyoffice-new-xlsx",
                 fileType: "xlsx",
                 actionHandler: function (name) {
-                    OCA.Onlyoffice.CreateFile(name + ".xlsx", fileList);
+                    if (!$("#isPublic").val() && OCA.Onlyoffice.TemplateExist("spreadsheet")) {
+                        OCA.Onlyoffice.OpenTemplatePicker(name, ".xlsx", "spreadsheet");
+                    } else {
+                        OCA.Onlyoffice.CreateFile(name + ".xlsx", fileList);
+                    }
                 }
             });
 
@@ -301,9 +287,17 @@
                 iconClass: "icon-onlyoffice-new-pptx",
                 fileType: "pptx",
                 actionHandler: function (name) {
-                    OCA.Onlyoffice.CreateFile(name + ".pptx", fileList);
+                    if (!$("#isPublic").val() && OCA.Onlyoffice.TemplateExist("presentation")) {
+                        OCA.Onlyoffice.OpenTemplatePicker(name, ".pptx", "presentation");
+                    } else {
+                        OCA.Onlyoffice.CreateFile(name + ".pptx", fileList);
+                    }
                 }
             });
+
+            if (OCA.Onlyoffice.GetTemplates) {
+                OCA.Onlyoffice.GetTemplates();
+            }
         }
     };
 
@@ -324,6 +318,12 @@
     OCA.Onlyoffice.bindVersionClick = function () {
         OCA.Onlyoffice.unbindVersionClick();
         $(document).on("click.onlyoffice-version", "#versionsTabView .downloadVersion", function() {
+            var ext = $(this).attr("download").split(".").pop();
+            if (!OCA.Onlyoffice.setting.formats[ext]
+                || !OCA.Onlyoffice.setting.formats[ext].def) {
+                return true;
+            }
+
             var versionNodes = $("#versionsTabView ul.versions>li");
             var versionNode = $(this).closest("#versionsTabView ul.versions>li")[0];