+++ js/viewer.js
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2020
+ * (c) Copyright Ascensio System SIA 2021
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -98,7 +98,7 @@
 
                 OCA.Viewer.registerHandler({
                     id: OCA.Onlyoffice.AppName,
-                    group: "documents",
+                    group: null,
                     mimes: mimes,
                     component: OnlyofficeViewerVue
                 })